sample-app/requirements.txt
# このファイルは参考用です
# 実際の依存関係はプロジェクトルートのpyproject.tomlで管理されています
# uvを使用してください: uv sync

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
Pillow==11.1.0
waitress==3.0.0
